@@ -2,6 +2,7 @@
   "dependencies": {
     "@aws-amplify/auth": "^3.4.16",
     "aws-lambda-cors": "0.0.2",
+    "aws-sdk": "^2.824.0",
     "crypto": "^1.0.1",
     "serverless": "^2.16.1",
     "serverless-cloudformation-sub-variables": "^0.2.1",
@@ -17,17 +18,23 @@
     "deploy:gateway": "util/build_swagger_yml.sh; sls deploy -v --force",
     "build:swagger-yaml": "util/build_swagger_yml.sh",
     "confirmAdmin": "util/confirm_admin.sh",
-    "openSwaggerUI": "util/open_swagger-ui.sh"
+    "openSwaggerUI": "util/open_swagger-ui.sh",
+    "prepublish": "tsc",
+    "dev": "webpack --mode development",
+    "build": "webpack --mode production"
   },
   "name": "remote-patient-monitoring",
   "devDependencies": {
     "@types/swagger-ui-dist": "^3.30.0",
+    "@types/uuid": "^8.3.0",
     "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^5.0.1",
     "html-webpack-plugin": "^4.5.0",
+    "js2ts": "^1.1.0",
     "light-server": "^2.9.1",
     "npm-run-all": "^4.1.5",
     "serverless-layers": "^2.3.3",
+    "serverless-plugin-typescript": "^1.1.9",
     "serverless-s3-deploy": "^0.9.0",
     "serverless-s3-remover": "^0.6.0",
     "serverless-scriptable-plugin": "^1.1.0",
@@ -35,7 +42,7 @@
     "swagger-cli": "^4.0.4",
     "swagger-merger": "^1.4.3",
     "ts-loader": "^8.0.13",
-    "typescript": "^4.1.3",
+    "typescript": "^4.2.0-dev.20210108",
     "webpack": "^4.44.2",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.1"
